requirements.txt
numpy
biopython
matplotlib
